some old updates

+++ MessageService/src/main/java/com/ttw/messageservice/repository/model/MessageRepository.java
@@ -3,5 +3,5 @@ package com.ttw.messageservice.repository.model;
 import com.ttw.messageservice.entity.Message;
 import com.ttw.messageservice.repository.BaseRepository;
 
-public interface MessageRepository extends BaseRepository<Message,Long> {
+public interface MessageRepository extends BaseRepository<Message,Integer> {
 }
